requirements.txt
Django==2.2.6
djangorestframework==3.10.3
pyasn1==0.4.7
pyasn1-modules==0.2.7
python-ldap==3.2.0
pytz==2019.3
sqlparse==0.3.0
